feat: make per a lib crate
expose account positions and other useful items
