+++ auction-server/src/auction/mod.rs
@@ -1,5 +1,4 @@
-mod repository;
-
 pub mod api;
 pub mod entities;
+pub mod repository;
 pub mod service;
